feat: add inventory controller
